gestion des notes

+++ app/Http/Controllers/NotesController.php
@@ -8,7 +8,11 @@ use App\Entree ;
 use App\Dossier ;
 use DB;
 use App\Note;
-use Auth; 
+use App\Note_his;
+use App\User;
+use App\EnvoyerNote;
+use App\EnvoyerNote_his;
+use auth; 
 use Illuminate\Routing\UrlGenerator;
 use URL;
 
@@ -28,55 +32,157 @@ class NotesController extends Controller
      * @return \Illuminate\Http\Response
      */
 
-
-    public function getNotesAjax ()
+     public function getNotesEnvoyeesAjax ()
     {
-         $dtc = (new \DateTime())->modify('-1 Hour')->format('Y-m-d H:i:s');
-         $notes=Note::where('date_rappel','<=', $dtc)->where('user_id', Auth::user()->id)->where('affichee','!=', 1)->get();
+
+      $dtc = (new \DateTime())->format('Y-m-d H:i:s');
+      $noteEnvoyee=EnvoyerNote::where('util_affecte',auth::user()->id)->where('date_rappel','<=', $dtc)->orderBy('date_rappel', 'asc')->orderBy('date_affectation', 'asc')->first();
+       $output='';
+
+
+   if( $noteEnvoyee){
+       $note=Note::where('id',$noteEnvoyee->note_id)->first();
+       $user=User::where('id',$note->emetteur_id)->first();
+    
+        if($noteEnvoyee->forceDelete())
+        {
+
+        $output='La note :  '.$note->titre.' | '.$note->contenu.', est vous a envoyée  par :'.$user->name.' '.$user->lastname;
+        }
+        else
+        {
+           $noteEnvoyee->forceDelete();
+           $output='La note : '.$note->titre.' | '.$note->contenu.' est envoyée à vous par :'.$user->name.' '.$user->lastname;
+
+        }
+
+      //Note::where('id',$note->id)->update(['affichee'=>1]);
+   
+     } 
+
+     return $output;
+   
+
+    }
+
+
+    public function getNotesReporteesAjax ()
+    {
+         $dtc = (new \DateTime())->format('Y-m-d H:i:s');
+         $noteReportee=Note::where('date_rappel','<=', $dtc)->where('statut','reportee')->where('user_id', auth::user()->id)->orderBy('date_rappel', 'asc')->first();
+
+         // vérifier que ce note n'est pas dans la table envoyerNote 
+         
 
      $output='';
 
-   if(count($notes)!=0){
-     foreach ($notes as $note) {
-       //$output.='<div>'. $note->id.'</div>';
-     $output.='<div class="row" style="padding: 0px; margin:0px" >'; 
+   if($noteReportee){
 
-    $output.=' <div class="col-md-10">
-        
-        <div class="panel panel-default">
-        <div class="panel-heading" style=" background-color: #00BFFF">
- 
-           <h4 class="panel-title">';
-              $output.=' <a data-toggle="collapse" href="#collapse'.$note->id.'"> '. $note->titre .'</a>';
-           $output.='</h4>
-        </div>';
+       $en=EnvoyerNote::where('note_id',$noteReportee->id)->first();
 
-      $output.='<div id="collapse'.$note->id.'" class="panel-collapse collapse in">';
-           $output.=' <ul class="list-group" style="padding:0px; margin:0px">';
-           
-             $output.='<li class="list-group-item"><a  href="#">'.$note->contenu.'</a></li>';
-          
-         $output.=' </ul>
+      if(!$en)
+      {
+       $output='Une nouvelle note est activée : '.$noteReportee->titre.' | '.$noteReportee->contenu;
+      }
 
-      </div>                                        
-    </div>
 
-  </div>
-
-    <div class="col-md-2">
+      $noteReportee->update(['statut'=>'active']);
    
-    </div>
+     } 
 
-
-  </div>';
-
-
-      Note::where('id',$note->id)->update(['affichee'=>1]);
-     }
- } 
-
-     echo $output;
+     return $output;
    
+
+    }
+
+     public function store(Request $request)
+    {
+
+
+      //titreNote
+      //descripNote
+      //daterappelNote
+     
+      
+      $user='';
+      $user_destin='';
+
+    if (!$request->get('EnvoyerNoteId') )
+    {
+       $user='moi';
+       $user_destin=auth::user()->id;
+    }
+    else
+    {
+
+         if( $request->get('EnvoyerNoteId')== auth::user()->id)
+         {
+           $user='moi';
+           $user_destin=auth::user()->id;
+         }
+         else
+         {
+           $user='autre';
+           $user_destin=trim($request->get('EnvoyerNoteId'));
+         }
+    }
+    //return   $user;
+
+
+        $format="Y-m-d\TH:i";
+        $dtc = (new \DateTime())->format('Y-m-d\TH:i');       
+        $dateSys  = \DateTime::createFromFormat($format, $dtc);
+        $dateNote  = \DateTime::createFromFormat($format, $request->get('daterappelNote'));          
+
+       if($dateNote>$dateSys)
+       {       
+        $statut='reportee';        
+       }
+       else
+       {
+        $statut='active';
+       }
+
+
+
+     $note = new Note([
+
+        'titre'=>trim($request->get('titreNote')),
+        'contenu'=>trim($request->get('descripNote')), 
+        'statut'=>$statut,
+        'date_rappel'=>trim($request->get('daterappelNote')),
+        'user_id'=>$user_destin,
+        'emetteur_id'=> auth::user()->id,
+        'originUser_id'=>auth::user()->id
+
+      ]);
+
+    $note->save();
+
+
+      if($user=='autre')// envoie note
+      {
+         $noteenvoie= new EnvoyerNote([
+         'util_affecteur'=>auth::user()->id, 
+         'util_affecte'=>$user_destin,
+         'note_id'=>$note->id,
+         'date_affectation'=> $dtc,        
+         'date_rappel'=>$request->get('daterappelNote'),
+         'statut'=>$statut
+
+         ]);
+         if($noteenvoie->save())
+         {
+          $envoyehis=new EnvoyerNote_his($noteenvoie->toArray());
+          $envoyehis->save();
+         }
+
+
+
+      }
+
+      return $dtc;
+
 
     }
 
@@ -87,7 +193,7 @@ class NotesController extends Controller
         $burl = URL::to("/");
 
 
-         $dtc = (new \DateTime())->modify('-1 Hour')->format('Y-m-d H:i:s');
+         $dtc = (new \DateTime())->format('Y-m-d H:i:s');
          $note=Note::where('date_rappel','<=', $dtc)->where('user_id', Auth::user()->id)->where('affichee','!=', 1)
          ->orderBy('date_rappel', 'asc')->first();
 
@@ -181,7 +287,7 @@ class NotesController extends Controller
    {
       $note = Note::find($id);
      // dd($note);
-      $note->update(['date_rappel'=>$request->get('daterappelNote'),'affichee'=>0]);
+      $note->update(['date_rappel'=>$request->get('daterappelNote'),'statut'=>'reportee']);
       return back();
 
 
@@ -192,7 +298,14 @@ class NotesController extends Controller
 
       
      $note = Note::find($id);
-     $note->delete();
+     $nh= new Note_his($note->toArray());
+
+     if($nh->save())
+     {
+        $nh->update(['note_id'=>$note->id]);
+        $note->delete();
+     }
+    
      return back();
 
 
@@ -236,25 +349,7 @@ class NotesController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
-    {
 
-        //dd($request->all());
-        $note = new Note([
-             'titre' =>trim( $request->get('titre')),
-             'contenu' => trim($request->get('descrip')),
-             'date_rappel'=> trim($request->get('daterappel')),
-             'affichee'=>0,
-             'user_id'=>auth::user()->id
-
-        ]);
-
-        $note->save();
-
-        return back();
-        /*return redirect('/notes')->with('success', ' ajouté avec succès');*/
-
-    }
 
     public function saving(Request $request)
     {
@@ -348,4 +443,127 @@ class NotesController extends Controller
         $note->delete();
 
         return redirect('/notes')->with('success', '  Supprimé  ');    }
+
+public function getAjaxUsersNote($idnote)
+{
+    $output='';
+
+     $note=Note::where('id',$idnote)->first();
+
+     $output.='<form  id="idFormUsersNote" method="get" action="">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h5 class="modal-title" id="exampleModal2">Envoi d\'une note </h5>
+
+            </div>
+            <div class="modal-body">
+                <div class="card-body">
+
+                    <div class="form-group">'.
+                                                
+                            csrf_field() .'
+                          
+                            <input id="EnvoyerNoteid" name="EnvoyerNoteid" type="hidden" value="'.$idnote.'">
+
+                            <input id="affecteurNote" name="affecteurNote" type="hidden" value="'. Auth::user()->id.'">
+                           
+
+                            <div class="form-group " >
+                                <div class=" row  ">
+                                    <div class="form-group mar-20">
+                                        <label for="agent" class="control-label" style="padding-right: 20px">Agent </label>
+                                        <select id="NoteAgent" name="NoteAgent" class="form-control select2" style="width: 230px">
+                                            <option value="Select">Selectionner</option>';
+                                              $agents = User::get(); 
+                                              $agentname='';
+                                                foreach ($agents as $agt){
+                                                 if (!empty ($agentname)) { 
+                                                 if ($agentname["id"] == $agt["id"]) {
+                                               $output.=' <option value="'. $agt["id"] .'" selected >'. $agt["name"] .'</option>';
+                                                }
+                                                else
+                                                {
+                                                 $output.=' <option value="'.$agt["id"] .'" >'. $agt["name"] .'</option>';
+                                                }
+                                               
+                                                
+                                                
+                                                }
+                                                else
+                                                  { $output.= '<option value="'.$agt["id"] .'" >'.$agt["name"].'</option>';}
+                                                
+                                               }   
+                                       $output.= ' </select>
+                                    </div>
+                                </div>
+                            </div>
+                      
+
+                    </div>
+
+
+                </div>
+
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn btn-secondary" data-dismiss="modal">Annuler</button>
+                <button type="button" id="BoutonEnvoyerNote" class="BoutonEnvoyerNote btn btn-secondary">Envoyer la note</button>
+            </div>
+        </div>
+          </form>';
+
+          return $output;
+
+}
+
+      public function EnvoyerNote (Request $request)
+      {
+        
+            /*"EnvoyerNoteid" => "envoyer8"
+              "affecteurNote" => "9"
+              "NoteAgent" => "Select"
+            "EnvoyerNoteid" => "8"
+              "affecteurNote" => "9"
+              "NoteAgent" => "Select"
+
+             */
+       $output='';
+        $format="Y-m-d\TH:i";
+        $dtc = (new \DateTime())->format('Y-m-d\TH:i');       
+        $dateSys  = \DateTime::createFromFormat($format, $dtc);
+       // $dateNote  = \DateTime::createFromFormat($format, $request->get('daterappelNote')); 
+
+      if($request->get('NoteAgent')!= 'Select' && $request->get('NoteAgent') )// envoie note
+      {
+         $noteenvoie= new EnvoyerNote([
+         'util_affecteur'=>auth::user()->id, 
+         'util_affecte'=>trim($request->get('NoteAgent')),
+         'note_id'=>trim($request->get('EnvoyerNoteid')),
+         'date_affectation'=> $dtc,        
+         'date_rappel'=>$dtc,
+         'statut'=>'active'
+
+         ]);
+         if($noteenvoie->save())
+         {
+          $envoyehis=new EnvoyerNote_his($noteenvoie->toArray());
+          $envoyehis->save();
+         }
+
+         $c=Note::where('id',$request->get('EnvoyerNoteid'))->first();
+         $c->update(['user_id'=>$request->get('NoteAgent')]);
+         $c->update(['emetteur_id'=>auth::user()->id]);
+
+            $output="la note est envoyée";
+
+      }
+      else
+      {
+        $output="Vous devez sélectionner un utilisateur";
+      }
+
+     return  $output;
+
+    }
+
 }
